# apps/api/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lumina API",
    "description": "Lumina 营销智能助手统一 API。涵盖意图识别、Skill 调用、Marketing Orchestra 编排、Token 用量查询、Demo 工作台等能力层接口。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/debug/chat/memory": {
      "get": {
        "tags": [
          "debug-chat"
        ],
        "summary": "查询 debug 对话记忆",
        "description": "查询指定 debug 对话的记忆列表。",
        "operationId": "get_memory_api_v1_debug_chat_memory_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "用户唯一标识",
              "title": "User Id"
            },
            "description": "用户唯一标识"
          },
          {
            "name": "conversation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "对话唯一标识",
              "title": "Conversation Id"
            },
            "description": "对话唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "debug-chat"
        ],
        "summary": "清除 debug 对话记忆",
        "description": "清除指定 debug 对话的全部记忆。",
        "operationId": "delete_memory_api_v1_debug_chat_memory_delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "用户唯一标识",
              "title": "User Id"
            },
            "description": "用户唯一标识"
          },
          {
            "name": "conversation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "对话唯一标识",
              "title": "Conversation Id"
            },
            "description": "对话唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/chat/stream": {
      "post": {
        "tags": [
          "debug-chat"
        ],
        "summary": "Debug Chat 流式对话",
        "description": "Debug Chat SSE 流式对话入口，支持 system_chat 走 MarketingOrchestra 编排。",
        "operationId": "chat_stream_api_v1_debug_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatStreamBody": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability",
            "description": "system_chat | content_direction_ranking | ..."
          },
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "description": "用户唯一标识",
            "default": "debug-user"
          },
          "conversation_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Conversation Id",
            "description": "对话唯一标识",
            "default": "debug-conv"
          },
          "message": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Message",
            "description": "用户当前输入"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform",
            "description": "可选：xiaohongshu / douyin 等"
          },
          "hub_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hub Context",
            "description": "与 POST /api/v1/marketing/hub 的 context 一致，传入 MarketingOrchestra（仅 system_chat 使用）"
          },
          "stream_format": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Format",
            "description": "1=v1 SSE；2=v2（仅 system_chat）。缺省可读 X-Lumina-Stream-Format"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "message"
        ],
        "title": "ChatStreamBody"
      },
...
      "DeleteMemoryResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "description": "操作是否成功"
          },
          "cleared": {
            "type": "boolean",
            "title": "Cleared",
            "description": "是否已清除"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "cleared"
        ],
        "title": "DeleteMemoryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "用户唯一标识"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "description": "对话唯一标识"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "消息数量"
          },
          "messages": {
            "items": {},
            "type": "array",
            "title": "Messages",
            "description": "消息列表"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "conversation_id",
          "count",
          "messages"
        ],
        "title": "MemoryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}